[🔨refactor] 상세페이지 섹션별로 컴포넌트 분리
상세페이지의 가독성을 높히기 위해 우선 5개의 섹션으로 분리했다
