+++ src/app/productdetail/_sections/RecommendedProductsSection.tsx
@@ -1,0 +1,7 @@
+export default function RecommendedProductsSection({ ref }) {
+  return (
+    <section ref={ref} className="h-[600px] w-full border border-pink-700">
+      <h3 className="text-48 font-bold">추천 상품</h3>
+    </section>
+  );
+}
