+++ src/app/productdetail/_sections/_index.ts
@@ -1,0 +1,5 @@
+export { default as ProductInfo } from "./ProductInfo";
+export { default as DetailNavigation } from "./DetailNavigation";
+export { default as Review } from "./Review";
+export { default as RecommendedProductsSection } from "./RecommendedProductsSection";
+export { default as DetailImage } from "./DetailImage";
